tty PICK_NONE menus
When testing the menu/incomplete map situation I noticed that <return>
didn't work to dismiss the "list autopickup exceptions" menu.  <space>
or <escape> was required.  That was clearly intentional but doesn't
seem reasonable.  Make <return> behave the same for PICK_NONE as it
does for other menu modes in tty and as it does for other interfaces.